Implements audio processing module in C++
This commit introduces a C++ audio processing module with Python bindings for improved performance.

It includes functionalities for:
- Calculating audio features like clarity, definition, RMS, and SNR.
- Augmenting reverberation time (TR) of impulse responses.
- Estimating and augmenting direct-to-reverberant ratio (DRR).
- Filtering audio signals using an octave filter bank.
- Calculating T30 using the Lundeby method.

The C++ implementation offers significant speed improvements compared to pure Python implementations. A `build.py` script is included to handle the compilation process, and pybind11 is used to create the Python bindings.

Also, updates the .gitignore to exlude .vscode folder.
A .gitmodules file is added for the pybind11 submodule